test: move frontend tests under __tests__
