+++ frontend/app/src/app/__tests__/AppProviders.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import { AppProviders } from './AppProviders';
+import { AppProviders } from '../AppProviders';
 
 describe('AppProviders', () => {
   it('renders children inside the application provider stack', () => {
