+++ frontend/app/src/app/__tests__/router.test.tsx
@@ -2,9 +2,12 @@ import { screen } from '@testing-library/react';
 import { RouterProvider } from '@tanstack/react-router';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import { renderWithAppProviders, resetFrontendTestState } from '../test/render';
-import { createAppRouter } from './router';
-import { routePath } from './routePaths';
+import {
+  renderWithAppProviders,
+  resetFrontendTestState,
+} from '../../test/render';
+import { createAppRouter } from '../router';
+import { routePath } from '../routePaths';
 
 describe('router', () => {
   beforeEach(() => {
