+++ frontend/app/src/routes/__tests__/LoginPage.test.tsx
@@ -1,8 +1,11 @@
 import { screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it } from 'vitest';
 
-import { renderWithAppProviders, resetFrontendTestState } from '../test/render';
-import { LoginPage } from './LoginPage';
+import {
+  renderWithAppProviders,
+  resetFrontendTestState,
+} from '../../test/render';
+import { LoginPage } from '../LoginPage';
 
 describe('LoginPage', () => {
   beforeEach(() => {
